+++ test/species_cache/test_species_cache.py
@@ -6,14 +6,14 @@ from ..mocks import mock_make_search_request
 
 
 class TestSpeciesCache:
-    def test_species_cacheby_tvk(self, client: TestClient, mocker):
+    def test_species_cache_by_tvk(self, client: TestClient, mocker):
         # Mock the Indicia warehouse.
         mocker.patch(
             'app.species.indicia.make_search_request',
             mock_make_search_request
         )
 
-        # Test cache is empty.
+        # Test cache is table is empty.
         response = client.get("/species/cache/count")
         assert response.status_code == 200
         result = response.json()
@@ -32,13 +32,13 @@ class TestSpeciesCache:
         assert taxon['preferred_tvk'] == "NBNSYS0000008319"
         assert taxon['organism_key'] == "NBNORG0000010513"
 
-        # Test cache is not empty.
+        # Test cache table is not empty.
         response = client.get("/species/cache/count")
         assert response.status_code == 200
         result = response.json()
         assert result['count'] == 1
 
-        # Request cache entry by id.
+        # Request cache table entry by id.
         response = client.get("/species/cache/1")
         assert response.status_code == 200
         taxon = response.json()
@@ -51,13 +51,13 @@ class TestSpeciesCache:
         assert taxon['preferred_tvk'] == "NBNSYS0000008319"
         assert taxon['organism_key'] == "NBNORG0000010513"
 
-        # Delete cache entry by id.
+        # Delete cache table entry by id.
         response = client.delete("/species/cache/1")
         assert response.status_code == 200
         result = response.json()
         assert result['ok'] is True
 
-        # Test cache is empty.
+        # Test cache table is empty.
         response = client.get("/species/cache/count")
         assert response.status_code == 200
         result = response.json()